Add the logo image to the page
